requirements.txt
# K-Line API 与训练依赖（从 environment.yml 提取）
# 建议使用 Python 3.11+
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==2.1.1
pydantic==2.8.2
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
tensorboard>=2.13.0
faiss-cpu>=1.7.4
yfinance>=0.2.18
plotly>=5.15.0
# CLIP（对比学习，训练 step4 需要）
git+https://github.com/openai/CLIP.git
